Move ensureFundsFor and related helpers into /lib/tank
